requirements.txt
wikipedia
python-dotenv
discord
wolframalpha
requests
